starts moving project to SPM by adding two packages, `Common` and `NSBrazilLib`
