--- NSBrazilConf/Extensions/Bundle+NSBrazilConf.swift
@@ -1,12 +1,0 @@
-
-import Foundation
-
-private final class _StubForBundleInit { }
-
-extension Bundle {
-
-    static var nsbrazilConf: Bundle {
-        return Bundle(for: _StubForBundleInit.self)
-    }
-
-}
