+++ NSBrazilConf/String+Capitalize.swift
@@ -1,12 +1,7 @@
-
-import UIKit
-
-extension String {
-
+public extension String {
     func capitalizedWord()-> String {
         let list = self.components(separatedBy: .whitespaces)
         let capitalized: [String] = list.map { $0.count <= 2 ? $0.lowercased()  : $0 }
         return capitalized.joined(separator: " ")
     }
-    
 }
